requirements.txt
python-vlc
pyyaml
mutagen
fastapi
uvicorn[standard]
pygame-ce
pylast
